Move GAUSS_EFF_AREA2FWHM_SQ to consts
